backend/requirements.txt
google-adk
anyio
python-dotenv
pydantic
fastapi 
uvicorn